+++ rust/candid/tests/assets/ok/example.rs
@@ -1,7 +1,7 @@
 // This is an experimental feature to generate Rust binding from Candid.
 // You may want to manually adjust some of the types.
-use candid::{self, CandidType, Deserialize};
-use ic_cdk::api::call::CallResult;
+use candid::{self, CandidType, Deserialize, Principal};
+use ic_cdk::api::call::CallResult as Result;
 
 #[derive(CandidType, Deserialize)]
 pub struct node { head: candid::Nat, tail: Box<list> }
@@ -9,7 +9,7 @@ pub struct node { head: candid::Nat, tail: Box<list> }
 #[derive(CandidType, Deserialize)]
 pub struct list(Option<node>);
 
-pub type my_type = candid::Principal;
+pub type my_type = Principal;
 #[derive(CandidType, Deserialize)]
 pub struct List_inner { head: candid::Int, tail: Box<List> }
 
@@ -57,21 +57,21 @@ pub struct b (candid::Int,candid::Nat,);
 #[derive(CandidType, Deserialize)]
 pub enum a { a, b(b) }
 
-pub struct SERVICE(pub candid::Principal);
-impl SERVICE{
+pub struct SERVICE(pub Principal);
+impl SERVICE {
   pub async fn f(
     &self,
     arg0: list,
-    arg1: Vec<u8>,
+    arg1: serde_bytes::ByteBuf,
     arg2: Option<bool>,
-  ) -> CallResult<()> { ic_cdk::call(self.0, "f", (arg0,arg1,arg2,)).await }
+  ) -> Result<()> { ic_cdk::call(self.0, "f", (arg0,arg1,arg2,)).await }
   pub async fn g(
     &self,
     arg0: my_type,
     arg1: List,
     arg2: Option<List>,
     arg3: nested,
-  ) -> CallResult<(candid::Int,broker,)> {
+  ) -> Result<(candid::Int,broker,)> {
     ic_cdk::call(self.0, "g", (arg0,arg1,arg2,arg3,)).await
   }
   pub async fn h(
@@ -79,13 +79,12 @@ impl SERVICE{
     arg0: Vec<Option<String>>,
     arg1: h_arg1,
     arg2: Option<List>,
-  ) -> CallResult<(h_ret0,)> {
-    ic_cdk::call(self.0, "h", (arg0,arg1,arg2,)).await
+  ) -> Result<(h_ret0,)> { ic_cdk::call(self.0, "h", (arg0,arg1,arg2,)).await }
+  pub async fn i(&self, arg0: List, arg1: f_arg1) -> Result<(Option<List>,)> {
+    ic_cdk::call(self.0, "i", (arg0,arg1,)).await
   }
-  pub async fn i(&self, arg0: List, arg1: f_arg1) -> CallResult<
-    (Option<List>,)
-  > { ic_cdk::call(self.0, "i", (arg0,arg1,)).await }
-  pub async fn x(&self, arg0: a, arg1: b) -> CallResult<
-    (Option<a>,Option<b>,)
-  > { ic_cdk::call(self.0, "x", (arg0,arg1,)).await }
+  pub async fn x(&self, arg0: a, arg1: b) -> Result<(Option<a>,Option<b>,)> {
+    ic_cdk::call(self.0, "x", (arg0,arg1,)).await
+  }
 }
+pub const service: SERVICE = SERVICE(Principal::from_slice(&[])); // aaaaa-aa